version 3.0.0
 - add `-n` flag instead of having each command separately implement its own code to parse
   a generic integer argument.
 - allow `tfm S X Y` in addition to `tfm S T` and `tfm S T X Y`.
 - add `with(thing, tmp_val, block)` macro.
 - add `nstatus_t` enum type with values: N_UNUSED, N_USED, and N_INFINITE.
 - add `u64 n;` and `nstatus_t nstatus;` to the global `cfg` object.
 - combine run and nrun together and rename it to frun.
 - rename `run` to `frun`.
 - merge nsim into sim.
 - rename bwrn/burn to brun.
 - add a `show` command that is equivalent to stepping 0 times or transforming with the noop
   transform and a shift of (0, 0).
 - add `WRAPPER` macro to conditionally exclude dump, fold, cnt, and merg commands.
 - fix version message in some instances.
 - allow stepping an infinite number of times in the step command.
    - if the final state is indeterminate, it will still figure out which cells don't change
      within the loop, and will print those cells as normal.
 - add a `rand` command that prints random states
 - add `bus` to the help text.
 - use `else ifeq` in Makefile.
 - add radix-sort.nasm for radix sort instead of introsort when ADX is enabled.
 - rename print_state to _print_state1 and add _print_state2 for with an indeterminate mask.
   this version can print alive, dead, and indeterminate cells
 - make the out of bounds check in _gen_combos a compile-time condition

+++ include/summary.h
@@ -127,14 +127,14 @@ static char *sprintf_summary(char *buf) {
 
 		for (u32 i = 0; i < PERIOD_MAX; i++) {
 			unlikely_if (i == max_period) {
-				BUF_WRITE(buf, "\"%u\": %zu", i, data.periods[i]);
+				BUF_WRITE(buf, "\"%zu\": %zu", (u64) i, data.periods[i]);
 				break;
 			}
 
 			if (data.periods[i] == 0)
 				continue;
 
-			BUF_WRITE(buf, "\"%u\": %zu", i, data.periods[i]);
+			BUF_WRITE(buf, "\"%zu\": %zu", (u64) i, data.periods[i]);
 			BUF_WRITE(buf, ',', ' ');
 		}
 	} // end bare block
@@ -150,14 +150,14 @@ static char *sprintf_summary(char *buf) {
 
 		for (u32 i = 0; i < TRANSIENT_MAX; i++) {
 			unlikely_if (i == max_transient) {
-				BUF_WRITE(buf, "\"%u\": %zu", i, data.transients[i]);
+				BUF_WRITE(buf, "\"%zu\": %zu", (u64) i, data.transients[i]);
 				break;
 			}
 
 			if (data.transients[i] == 0)
 				continue;
 
-			BUF_WRITE(buf, "\"%u\": %zu", i, data.transients[i]);
+			BUF_WRITE(buf, "\"%zu\": %zu", (u64) i, data.transients[i]);
 			BUF_WRITE(buf, ',', ' ');
 		}
 	} // end bare block
